Move spotify client code to own package folder
