+++ src/packages/spotify/models/album.py
@@ -1,5 +1,5 @@
-from app.models.track import Track
-from app.models.artist import Artist
+from packages.spotify.models.track import Track
+from packages.spotify.models.artist import Artist
 from datetime import datetime
