+++ src/packages/spotify/models/artist.py
