+++ src/scripts/update_playlist_based_on_stats.py
@@ -7,7 +7,7 @@ from app.lib.music_util import MusicUtil
 from app.lib.my_music_lib import MyMusicLib
 from app.lib.playlist_updater import PlaylistUpdater
 from app.lib.playlist_stats import PlaylistStats
-from app.lib.spotify_client_wrapper import SpotifyClientWrapper
+from packages.spotify.spotify_client_wrapper import SpotifyClientWrapper
 
 
 def main():
