+++ src/tests/fixtures.py
@@ -1,9 +1,9 @@
-from app.models.album import Album
-from app.models.artist import Artist
-from app.models.audio_features import AudioFeatures
-from app.models.playlist import Playlist
-from app.models.track import Track
-from app.models.song_attribute_ranges import SongAttributeRanges
+from packages.spotify.models.album import Album
+from packages.spotify.models.artist import Artist
+from packages.spotify.models.audio_features import AudioFeatures
+from packages.spotify.models.playlist import Playlist
+from packages.spotify.models.track import Track
+from packages.spotify.models.song_attribute_ranges import SongAttributeRanges
 
 def mock_album(id="", genres=[], artists=[], name=""):
     return Album(name, id, [], artists, release_date="", num_tracks=0, genres=genres)
